finished lose and win screens with scroll implementation
