+++ Assets/Scripts/Win Lose Screen/TryAgainScript.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TryAgainScript : MonoBehaviour
+{
+    public void TryAgain()
+    {
+        GameManager.instance.LoadMainMenu();
+    }
+}
