--- Assets/Scripts/Win Lose Screen/WinLose.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class WinLose : MonoBehaviour
-{
-    public void TryAgain()
-    {
-        SceneManager.LoadScene("MainMenu");
-    }
-    public void DisplayDynastyStats()
-    {
-        List<string> DynastyStats = new List<string>(); 
-    }
-}
