--- Assets/Scripts/Win Lose Screen/WinLose.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 31434a6422862f84eb358bdd9c5ca651
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
